Start Forage object; Move AV to ImageRecognition;
The Forage object is going to implement the high level CPFA logic as well as communication with the arduino.  Currently it is very much a work in progress, with lots of comments and pseudocode.

Attempted to refactor all of the image recognition code out of MainController.  Hopefully this does not have negative consequences.  ImageRecognition is also in shambles right now.  There is a small amount of work needed to make it functional again, but this relies on Forage congealing a bit more.

Router and RouterServer changes are negligible.